Some refactoring to simplify spring projects (both spring-mvc and spring-boot).
