+++ java/spring_boot_helloworld/Dockerfile
@@ -4,8 +4,8 @@ MAINTAINER burrsutter@gmail.com
 
 RUN yum -y install java
 
-ADD target/springboot-0.1.0.jar .
+ADD target/springboot-helloworld.jar .
 
 EXPOSE 8080
 
-CMD java -jar springboot-0.1.0.jar
+CMD java -jar springboot-helloworld.jar
