+++ java/spring_boot_helloworld/pom.xml
@@ -4,9 +4,8 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.redhat.samples</groupId>
-    <artifactId>springboot</artifactId>
-    <version>0.1.0</version>
-
+    <artifactId>springboot-helloworld</artifactId>
+    <version>1.0.0-SNAPSHOT</version>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
@@ -25,8 +24,8 @@
         <java.version>1.8</java.version>
     </properties>
 
-
     <build>
+        <finalName>${project.artifactId}</finalName>
         <plugins>
             <plugin>
                 <groupId>org.springframework.boot</groupId>
