+++ java/spring_boot_helloworld/src/main/java/com/redhat/samples/springboot/domain/User.java
@@ -1,27 +1,34 @@
-package com.redhat.samples.springmvc.models;
+package com.redhat.samples.springboot.domain;
 
 public class User {
 	private String name;
+
 	private String address;
+
 	private int age;
+
 	public String getName() {
 		return name;
 	}
+
 	public void setName(String name) {
 		this.name = name;
 	}
+
 	public String getAddress() {
 		return address;
 	}
+
 	public void setAddress(String address) {
 		this.address = address;
 	}
+
 	public int getAge() {
 		return age;
 	}
+
 	public void setAge(int age) {
 		this.age = age;
 	}
-	
- 
- }
+
+}
