+++ java/tomcat_springmvc/src/main/java/com/redhat/samples/springmvc/Application.java
@@ -1,0 +1,12 @@
+package com.redhat.samples.springmvc;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+
+@Configuration
+@EnableWebMvc
+@ComponentScan
+public class Application {
+
+}
