+++ java/tomcat_springmvc/src/main/java/com/redhat/samples/springmvc/domain/User.java
@@ -1,39 +1,34 @@
-
-package com.redhat.samples.springboot.models;
+package com.redhat.samples.springmvc.domain;
 
 public class User {
 	private String name;
+
 	private String address;
+
 	private int age;
-	
-    public User() {
-	
-    }
-	
-	public User(String name, String address, int age) {
-		this.name = name;
-		this.address = address;
-		this.age = age;
-	}
-	
+
 	public String getName() {
 		return name;
 	}
+
 	public void setName(String name) {
 		this.name = name;
 	}
+
 	public String getAddress() {
 		return address;
 	}
+
 	public void setAddress(String address) {
 		this.address = address;
 	}
+
 	public int getAge() {
 		return age;
 	}
+
 	public void setAge(int age) {
 		this.age = age;
 	}
-	
- 
- }
+
+}
