--- java/tomcat_springmvc/src/main/webapp/WEB-INF/views/welcome.jsp
@@ -1,12 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>View associated with Welcome.java</title>
-</head>
-<body>
-	<h3> Hello ${user.name} </h3>
-	${user.address} <br/>
-	${user.age} <p/>
-</body>
-</html>
