--- java/tomcat_springmvc/src/main/webapp/index.jsp
@@ -1,15 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Index.jsp</title>
-</head>
-<body>
-Stuff Happens <br> <%= new java.util.Date() %>
-<%
-System.out.println("invoking index.jsp");
-%>
-</body>
-</html>
